Created new develop branch for adding code features while maintaining stable code in master branch; Added features and bug fixes
Fixed issue with data being flipped vertically. The PLC data is flipped vertically, so another vertical flip puts it right side up.

It is now possible to work on rectangular (non-square) input data.

Fixed file seperator characters in the live script to accomodate UNIX/Linux.

Added in a safety measure which prompts the user to select the AutoSaves directory again if it was incorrectly selected the first time.