Updated gradient calculation and plotting
The manual input of cut line vertices is now functional with the gridded data and the gradient of the gridded data. The next step is to make the automatic detection of the maximum gradient and associated placement of the cutline compatible with the gridded data created from the scattered interpolant.