Simplified code and revamped PlotPLC script
The PlotPLC script is now set up to serve as a demonstration of the syntax used to use PLCtools from a regular MATLAB script or the MATLAB Command Window. The code was simplified to make the process of setting cut lines, etc. more automatic / require fewer PLC object method calls to achieve the same results.