Bug fixes
Revised data load method so that the proper .mat file is loaded even if the gridded results are missing.

Regarding the issue with custom data text not working for the PlotPLC Live Script, I submitted a support request to Mathworks after discovering that the code to produce a figure with functioning custom data cursor text works fine with a regular .m file but not with a .mlx Live Script.